+++ src/common/index.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <table width="100%" cellpadding="0" cellspacing="0" style="padding: 40px 0; font-family: Arial, sans-serif;">
+    <tr>
+        <td align="center">
+        <table width="600" cellpadding="0" cellspacing="0" style="background-color: #005baa; padding: 30px; border-radius: 8px;">
+            <tr>
+            <td align="left" style="color: #ffffff;">
+                <p>Hola ${donorName},</p>
+                <h2 style="margin-top: 0; color: #ffffff;">Fundación - PFU Peerkals Fundation</h2>
+                <p style="margin-bottom: 20px; color: #ffffff;">
+                Gracias por tu valiosa contribución. Tu generosidad impulsa nuestro propósito de transformar vidas a través de la tecnología y la educación. 
+                <br /><br />
+                Adjuntamos a este correo el certificado oficial de tu donación.
+                </p>
+            </td>
+            </tr>
+        </table>
+        </td>
+    </tr>
+    <tr>
+        <td align="center" style="padding-top: 20px;">
+        <img src="https://cdn1.site-media.eu/images/0/17352501/PFU-Logo_blanco-GxBlRzv5qp7Lsp-cXOtV1g.png" alt="Logo Fundación PFU" width="120" style="display: block;" />
+        </td>
+    </tr>
+</table>
+</body>
+</html>
