+++ src/endpoints/signatureIntegrity.ts
@@ -66,9 +66,7 @@ export class SignatureIntegrity extends OpenAPIRoute{
 }
 
 async function generateSignatureIntegrity(currency: string, amount: string, uiid:string) {
-  //const fechaActual = new Date().toISOString();
-  //   "<Referencia><Monto><Moneda><FechaExpiracion><SecretoIntegridad>"
-  const rawValue=uiid.concat(amount).concat(currency).concat(Security.SECRET_KEY.toString());  
+  const rawValue=uiid.concat(amount).concat(currency).concat(Security.SECRET_KEY_TEST.toString());  
   const hash = hashSHA256(rawValue);
   return hash;
 }
